MAJOR rehaul of backend api - moved to blueprint-based api (will need to implement table check), get and edit functions for basic clinical info added, still requires add functions. Schema fixed and tested.
