+++ extensions.py
@@ -1,0 +1,2 @@
+from flask_mysqldb import MySQL, MySQLdb
+mysql = MySQL()
